+++ apps/tergite-mss/services/jobs/__init__.py
@@ -16,14 +16,17 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 import logging
-from typing import TYPE_CHECKING, List, Mapping, Optional, Tuple
+from pathlib import Path
+from typing import TYPE_CHECKING, List, Optional, Tuple
 from uuid import UUID
 
 from motor.motor_asyncio import AsyncIOMotorDatabase
 from pymongo import ReturnDocument
 
 from services.external.bcc import BccClient
+from settings import PRIVATE_KEY_FILE
 from utils import mongodb as mongodb_utils
+from utils.crypto import decrypt_message
 from utils.exc import NotFoundError
 
 from ..auth import Project
@@ -33,12 +36,19 @@ if TYPE_CHECKING:
     from ..auth.projects.database import ProjectDatabase
 
 
-async def get_one(db: AsyncIOMotorDatabase, job_id: UUID) -> Job:
+async def get_one(
+    db: AsyncIOMotorDatabase,
+    job_id: UUID,
+    is_token_decrypted: bool = False,
+    private_key_file: Path = PRIVATE_KEY_FILE,
+) -> Job:
     """Gets a job by job_id
 
     Args:
         db: the mongo database from where to get the job
         job_id: the `job_id` of the job to be returned
+        is_token_decrypted: whether the token returned is decrypted
+        private_key_file: the path to the private key
 
     Raises:
         utils.exc.NotFoundError: no matches for '{search_filter}'.
@@ -47,16 +57,23 @@ async def get_one(db: AsyncIOMotorDatabase, job_id: UUID) -> Job:
     Returns:
         the job as a dict
     """
-    return await mongodb_utils.find_one(
+    result = await mongodb_utils.find_one(
         db.jobs, {"job_id": str(job_id)}, schema=Job, dropped_fields=("_id",)
     )
+    if is_token_decrypted and result.access_token:
+        result.access_token = decrypt_message(
+            private_key_file=private_key_file, msg=result.access_token
+        )
+
+    return result
 
 
 async def create_job(
     db: AsyncIOMotorDatabase,
     bcc_client: BccClient,
     job: Job,
-    app_token: Optional[str] = None,
+    user_id: str,
+    request_id: str,
 ) -> CreatedJobResponse:
     """Creates a new job for the given backend
 
@@ -64,7 +81,8 @@ async def create_job(
         db: the mongo database from where to create the job
         bcc_client: the HTTP client for accessing BCC
         job: the job object to create
-        app_token: the associated with this new job. It is None if no auth is required
+        user_id: the unique identifier of the user
+        request_id: the unique identifier of the current request
 
     Returns:
         the created job details
@@ -76,7 +94,10 @@ async def create_job(
                 See :meth:`utils.http_clients.BccClient.save_credentials`
     """
     logging.info(f"Creating new job with id: {job.job_id}")
-    await bcc_client.save_credentials(job_id=job.job_id, app_token=f"{app_token}")
+    encrypted_token, token = await bcc_client.get_token(
+        job_id=job.job_id, user_id=user_id, request_id=request_id
+    )
+    job.access_token = encrypted_token
     await mongodb_utils.insert_one(collection=db.jobs, document=job.model_dump())
 
     upload_url = _without_special_docker_host_domain(f"{bcc_client.base_url}/jobs")
@@ -84,6 +105,7 @@ async def create_job(
     return {
         "job_id": job.job_id,
         "upload_url": upload_url,
+        "access_token": token,
     }
 
 
@@ -92,10 +114,13 @@ async def get_latest_many(
     filters: Optional[dict] = None,
     limit: Optional[int] = None,
     skip: int = 0,
-    exclude: Tuple[str] = (),
+    exclude: Tuple[str] = ("access_token",),
     sort: List[str] = (),
 ) -> List[Job]:
     """Retrieves the latest jobs up to the given limit
+
+    By default, the access_token is not returned when many jobs are requested for.
+    This is because such requests are usually not to submit the job but rather to view it.
 
     Args:
         db: the mongo database from where to get the jobs