+++ apps/tergite-mss/tests/_utils/env.py
@@ -20,16 +20,14 @@ import tomli
 from .fixtures import get_fixture_path
 
 TEST_MSS_CONFIG_FILE = get_fixture_path("config.test.toml")
-TEST_PROD_NO_AUTH_MSS_CONFIG_FILE = get_fixture_path(
-    "prod_config_with_no_auth.test.toml"
-)
-TEST_NO_AUTH_MSS_CONFIG_FILE = get_fixture_path("disabled_auth_config.test.toml")
 TEST_DISABLED_PUHURI_MSS_CONFIG_FILE = get_fixture_path(
     "disabled_puhuri_config.test.toml"
 )
 with Path(TEST_MSS_CONFIG_FILE).open(mode="rb") as _oauth_conf_file:
     TEST_APP_CONFIG = tomli.load(_oauth_conf_file)
 
+TEST_MSS_PUBLIC_KEY_PATH = get_fixture_path("mss_public_key.pem")
+TEST_MSS_PRIVATE_KEY_PATH = get_fixture_path("mss_private_key.pem")
 
 TEST_PUHURI_CONFIG_ENDPOINT = TEST_APP_CONFIG["auth"]["clients"][2][
     "openid_configuration_endpoint"
@@ -50,6 +48,7 @@ def setup_test_env():
     It should be run before any imports
     """
     environ["MSS_CONFIG_FILE"] = TEST_MSS_CONFIG_FILE
+    environ["PRIVATE_KEY_FILE"] = TEST_MSS_PRIVATE_KEY_PATH
 
 
 class BackendConfDict(TypedDict):