+++ apps/tergite-mss/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project follows versions of format `{year}.{month}.{patch_number}`.
 
 ## [Unreleased]
 
+### Added
+
+- Added `access_token` field to the job schema. 
+  It is encrypted when in database but plain when requested for by anyone who has access to the job
+  (including other members of the project).
+
+### Changed
+
+- BREAKING: Changed the response to job creation `/jobs` POST endpoint, to return a JWT access token also
+- BREAKING: Changed the authentication with BCC to use private key signed signatures in headers
+- BREAKING: Removed the option to disable authentication as it is a requirement for BCC
+
 ## [2025.06.2] - 2025-06-17
 
 - No change